@@ -7,20 +7,27 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.6.10",
+    "@chakra-ui/descendant": "^3.0.3",
+    "@chakra-ui/icons": "^2.2.4",
+    "@chakra-ui/react": "^2.10.9",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "@react-google-maps/api": "^2.20.6",
+    "framer-motion": "^10.18.0",
     "path": "^0.12.7",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.5.0",
     "sass": "^1.43.4",
+    "tslib": "^2.4.0",
     "vite-tsconfig-paths": "^3.3.17"
   },
   "devDependencies": {
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/node": "^16.18.126",
+    "@types/react": "^18.3.23",
+    "@types/react-dom": "^18.3.7",
     "@vitejs/plugin-react": "^1.0.0",
-    "typescript": "^4.3.2",
+    "typescript": "^5.4.5",
     "vite": "^2.6.4"
   }
 }
