@@ -1,15 +1,32 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import './index.scss'
-import App from './App'
-import { ChakraProvider } from "@chakra-ui/react";
-import { theme } from './theme';
+// import React from 'react'
+// import ReactDOM from 'react-dom'
+// import './index.scss'
+// import App from './App'
+// import { ChakraProvider } from "@chakra-ui/react";
+// import { theme } from './theme';
 
-ReactDOM.render(
+// ReactDOM.render(
+//   <React.StrictMode>
+//     <ChakraProvider theme={theme}>
+//       <App />
+//     </ChakraProvider>
+//   </React.StrictMode>,
+//   document.getElementById('root')
+// ) 
+
+
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import { ChakraProvider } from '@chakra-ui/react';
+import App from './App';
+import { theme } from './theme';
+import './index.scss';
+
+const container = document.getElementById('root')!;
+createRoot(container).render(
   <React.StrictMode>
     <ChakraProvider theme={theme}>
       <App />
     </ChakraProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+  </React.StrictMode>
+);
